reservation_backend/requirements.txt
pytest-django
djangorestframework-simplejwt[crypto]
django-cors-headers
drf-spectacular
requests
django-filter
psycopg2-binary
gunicorn
python-dotenv
